Fixed the prime number & 1√root major bugs
Replaced the list of prime numbers by a list of numbers from 0 to 1500 (subject to be changed in the future), done this by a for loop.  Also fixed the 1√root  (where root is the argument parsed) bug by an if statement that checks the mod_s variable, returns  √root if mod_s is 1. 1√root is 1 * √root so same thing.