Added  x√root Major Feature
Added a  x√root where the  x variable represents the 4 and the root represents the 125 in the expression 4√125. Did this using the *root (optional argument in the spl function) that takes the numbers before and after the ! (using ! instead of √ because square root symbol (√) is not on the keyboard, so 4√125  =  4!125 and √125  =  !125). Mapped the root variable to the number after the ! and the multp to the number before the ! (in this case the 4 is mapped to multp and the 125 to root) and used the .split() method for the program to return a list (later used with indexes for the algorithm). Also took some time to order the code and modified the start & end constants to 100 000 instead of 5000, for the program to have a wider range of possible simplifications (this slowed the algorithm down by 0.1 seconds, not a big deal! )